refactor ItemServiceTests.
